Crate revm_interpreter

revm-interpreter

REVM Interpreter.

Re-exports

pub use gas::Gas;

pub use instructions::opcode;

pub use instructions::Instruction;

pub use instructions::OpCode;

pub use instructions::OPCODE_JUMPMAP;

Modules

analysis

gas

EVM gas calculation utilities.

instructions

EVM opcode implementations.

Macros

return_ok

return_revert

Structs

BytecodeLocked

An analyzed bytecode.

CallContext

Context of a runtime call.

CallInputs

Inputs for a call.

Contract

EVM contract information.

CreateInputs

Inputs for a create call.

DummyHost

A dummy Host implementation.

Interpreter

SelfDestructResult

Result of a call that resulted in a self destruct.

SharedMemory

A sequential memory shared between calls, which uses a Vec for internal representation. A SharedMemory instance should always be obtained using the new static method to ensure memory safety.

Stack

EVM stack.

Transfer

Transfer from source to target, with given value.

Enums

CallScheme

Call schemes.

CreateScheme

Create scheme.

InstructionResult

SuccessOrHalt

Constants

MAX_CODE_SIZE

EIP-170: Contract code size limit

MAX_INITCODE_SIZE

EIP-3860: Limit and meter initcode

STACK_LIMIT

EVM interpreter stack limit.

Traits

Host

EVM context host.

Functions

next_multiple_of_32

Rounds up x to the closest multiple of 32. If x % 32 == 0 then x is returned.