Module revm_interpreter::gas

Structs

Gas

Constants

ACCESS_LIST_ADDRESS

ACCESS_LIST_STORAGE_KEY

BASE

BLOCKHASH

CALLVALUE

CALL_STIPEND

CODEDEPOSIT

COLD_ACCOUNT_ACCESS_COST

COLD_SLOAD_COST

COPY

CREATE

EXP

HIGH

INITCODE_WORD_COST

EIP-3860 : Limit and meter initcode

JUMPDEST

LOG

LOGDATA

LOGTOPIC

LOW

MEMORY

MID

NEWACCOUNT

REFUND_SSTORE_CLEARS

SELFDESTRUCT

SHA3

SHA3WORD

SSTORE_RESET

SSTORE_SET

TRANSACTION_NON_ZERO_DATA_FRONTIER

TRANSACTION_NON_ZERO_DATA_INIT

TRANSACTION_ZERO_DATA

VERYLOW

WARM_STORAGE_READ_COST

ZERO

Functions

account_access_gas

call_cost

create2_cost

exp_cost

extcodecopy_cost

hot_cold_cost

initcode_cost

EIP-3860: Limit and meter initcode apply extra gas cost of 2 for every 32-byte chunk of initcode Can’t overflow as initcode length is assumed to be checked

log_cost

memory_gas

selfdestruct_cost

sha3_cost

sload_cost

sstore_cost

sstore_refund

verylowcopy_cost