revm_interpreter::host

Trait Host

pub trait Host {

    // Required methods
    fn basefee(&self) -> U256;
    fn blob_gasprice(&self) -> U256;
    fn gas_limit(&self) -> U256;
    fn difficulty(&self) -> U256;
    fn prevrandao(&self) -> Option<U256>;
    fn block_number(&self) -> u64;
    fn timestamp(&self) -> U256;
    fn beneficiary(&self) -> Address;
    fn chain_id(&self) -> U256;
    fn effective_gas_price(&self) -> U256;
    fn caller(&self) -> Address;
    fn blob_hash(&self, number: usize) -> Option<U256>;
    fn initcode_by_hash(&mut self, hash: B256) -> Option<Bytes>;
    fn max_initcode_size(&self) -> usize;
    fn block_hash(&mut self, number: u64) -> Option<B256>;
    fn selfdestruct(
        &mut self,
        address: Address,
        target: Address,
    ) -> Option<StateLoad<SelfDestructResult>>;
    fn log(&mut self, log: Log);
    fn sstore(
        &mut self,
        address: Address,
        key: StorageKey,
        value: StorageValue,
    ) -> Option<StateLoad<SStoreResult>>;
    fn sload(
        &mut self,
        address: Address,
        key: StorageKey,
    ) -> Option<StateLoad<StorageValue>>;
    fn tstore(&mut self, address: Address, key: StorageKey, value: StorageValue);
    fn tload(&mut self, address: Address, key: StorageKey) -> StorageValue;
    fn balance(&mut self, address: Address) -> Option<StateLoad<U256>>;
    fn load_account_delegated(
        &mut self,
        address: Address,
    ) -> Option<StateLoad<AccountLoad>>;
    fn load_account_code(
        &mut self,
        address: Address,
    ) -> Option<StateLoad<Bytes>>;
    fn load_account_code_hash(
        &mut self,
        address: Address,
    ) -> Option<StateLoad<B256>>;

}

Host trait with all methods that are needed by the Interpreter.

This trait is implemented for all types that have ContextTr trait.

There are few groups of functions which are Block, Transaction, Config, Database and Journal functions.

Required Methods

fn basefee(&self) -> U256

Block basefee, calls ContextTr::block().basefee()

fn blob_gasprice(&self) -> U256

Block blob gasprice, calls ContextTr::block().blob_gasprice()

fn gas_limit(&self) -> U256

Block gas limit, calls ContextTr::block().gas_limit()

fn difficulty(&self) -> U256

Block difficulty, calls ContextTr::block().difficulty()

fn prevrandao(&self) -> Option<U256>

Block prevrandao, calls ContextTr::block().prevrandao()

fn block_number(&self) -> u64

Block number, calls ContextTr::block().number()

fn timestamp(&self) -> U256

Block timestamp, calls ContextTr::block().timestamp()

fn beneficiary(&self) -> Address

Block beneficiary, calls ContextTr::block().beneficiary()

fn chain_id(&self) -> U256

Chain id, calls ContextTr::cfg().chain_id()

fn effective_gas_price(&self) -> U256

Transaction effective gas price, calls ContextTr::tx().effective_gas_price(basefee as u128)

fn caller(&self) -> Address

Transaction caller, calls ContextTr::tx().caller()

fn blob_hash(&self, number: usize) -> Option<U256>

Transaction blob hash, calls ContextTr::tx().blob_hash(number)

fn initcode_by_hash(&mut self, hash: B256) -> Option<Bytes>

Initcodes mapped to the hash.

fn max_initcode_size(&self) -> usize

Max initcode size, calls ContextTr::cfg().max_code_size().saturating_mul(2)

fn block_hash(&mut self, number: u64) -> Option<B256>

Block hash, calls ContextTr::journal().db().block_hash(number)

fn selfdestruct( &mut self, address: Address, target: Address, ) -> Option<StateLoad<SelfDestructResult>>

Selfdestruct account, calls ContextTr::journal().selfdestruct(address, target)

fn log(&mut self, log: Log)

Log, calls ContextTr::journal().log(log)

fn sstore( &mut self, address: Address, key: StorageKey, value: StorageValue, ) -> Option<StateLoad<SStoreResult>>

Sstore, calls ContextTr::journal().sstore(address, key, value)

fn sload( &mut self, address: Address, key: StorageKey, ) -> Option<StateLoad<StorageValue>>

Sload, calls ContextTr::journal().sload(address, key)

fn tstore(&mut self, address: Address, key: StorageKey, value: StorageValue)

Tstore, calls ContextTr::journal().tstore(address, key, value)

fn tload(&mut self, address: Address, key: StorageKey) -> StorageValue

Tload, calls ContextTr::journal().tload(address, key)

fn balance(&mut self, address: Address) -> Option<StateLoad<U256>>

Balance, calls ContextTr::journal().load_account(address)

fn load_account_delegated( &mut self, address: Address, ) -> Option<StateLoad<AccountLoad>>

Load account delegated, calls ContextTr::journal().load_account_delegated(address)

fn load_account_code(&mut self, address: Address) -> Option<StateLoad<Bytes>>

Load account code, calls ContextTr::journal().load_account_code(address)

fn load_account_code_hash( &mut self, address: Address, ) -> Option<StateLoad<B256>>

Load account code hash, calls ContextTr::journal().code_hash(address)