Module revm_interpreter::interpreter_action

Structs

CallInputs

Inputs for a call.

CallOutcome

Represents the outcome of a call operation in a virtual machine.

CreateInputs

Inputs for a create call.

CreateOutcome

Represents the outcome of a create operation in an interpreter.

EOFCreateInputs

Inputs for EOF create call.

Enums

CallScheme

Call scheme.

CallValue

Call value.

CreateScheme

Create scheme.

EOFCreateKind

EOF create can be called from two places:

InterpreterAction