Struct revm_interpreter::opcode::OpCode

pub struct OpCode(/* private fields */);

An EVM opcode.

This is always a valid opcode, as declared in the opcode module or the OPCODE_INFO_JUMPTABLE constant.

Implementations

impl OpCode

pub const fn new(opcode: u8) -> Option<Self>

Instantiate a new opcode from a u8.

pub const fn is_jumpdest(&self) -> bool

Returns true if the opcode is a jump destination.

pub const fn is_jumpdest_by_op(opcode: u8) -> bool

Takes a u8 and returns true if it is a jump destination.

pub const fn is_jump(self) -> bool

Returns true if the opcode is a legacy jump instruction.

pub const fn is_jump_by_op(opcode: u8) -> bool

Takes a u8 and returns true if it is a jump instruction.

pub const fn is_push(self) -> bool

Returns true if the opcode is a PUSH instruction.

pub fn is_push_by_op(opcode: u8) -> bool

Takes a u8 and returns true if it is a push instruction.

pub unsafe fn new_unchecked(opcode: u8) -> Self

Instantiate a new opcode from a u8 without checking if it is valid.

Safety

All code using Opcode values assume that they are valid opcodes, so providing an invalid opcode may cause undefined behavior.

pub const fn as_str(self) -> &'static str

Returns the opcode as a string. This is the inverse of parse.

pub const fn name_by_op(opcode: u8) -> &'static str

Returns the opcode name.

pub const fn inputs(&self) -> u8

Returns the number of input stack elements.

pub const fn outputs(&self) -> u8

Returns the number of output stack elements.

pub const fn io_diff(&self) -> i16

Calculates the difference between the number of input and output stack elements.

pub const fn info_by_op(opcode: u8) -> Option<OpCodeInfo>

Returns the opcode information for the given opcode.

pub const fn info(&self) -> OpCodeInfo

Returns the opcode information.

pub const fn input_output(&self) -> (u8, u8)

Returns the number of both input and output stack elements.

Can be slightly faster that calling inputs and outputs separately.

pub const fn get(self) -> u8

Returns the opcode as a u8.

pub const fn modifies_memory(&self) -> bool

Returns true if the opcode modifies memory. https://bluealloy.github.io/revm/crates/interpreter/memory.html#opcodes https://github.com/crytic/evm-opcodes

impl OpCode

pub const STOP: Self = _

The 0x00 (“STOP”) opcode.

pub const ADD: Self = _

The 0x01 (“ADD”) opcode.

pub const MUL: Self = _

The 0x02 (“MUL”) opcode.

pub const SUB: Self = _

The 0x03 (“SUB”) opcode.

pub const DIV: Self = _

The 0x04 (“DIV”) opcode.

pub const SDIV: Self = _

The 0x05 (“SDIV”) opcode.

pub const MOD: Self = _

The 0x06 (“MOD”) opcode.

pub const SMOD: Self = _

The 0x07 (“SMOD”) opcode.

pub const ADDMOD: Self = _

The 0x08 (“ADDMOD”) opcode.

pub const MULMOD: Self = _

The 0x09 (“MULMOD”) opcode.

pub const EXP: Self = _

The 0x0A (“EXP”) opcode.

pub const SIGNEXTEND: Self = _

The 0x0B (“SIGNEXTEND”) opcode.

pub const LT: Self = _

The 0x10 (“LT”) opcode.

pub const GT: Self = _

The 0x11 (“GT”) opcode.

pub const SLT: Self = _

The 0x12 (“SLT”) opcode.

pub const SGT: Self = _

The 0x13 (“SGT”) opcode.

pub const EQ: Self = _

The 0x14 (“EQ”) opcode.

pub const ISZERO: Self = _

The 0x15 (“ISZERO”) opcode.

pub const AND: Self = _

The 0x16 (“AND”) opcode.

pub const OR: Self = _

The 0x17 (“OR”) opcode.

pub const XOR: Self = _

The 0x18 (“XOR”) opcode.

pub const NOT: Self = _

The 0x19 (“NOT”) opcode.

pub const BYTE: Self = _

The 0x1A (“BYTE”) opcode.

pub const SHL: Self = _

The 0x1B (“SHL”) opcode.

pub const SHR: Self = _

The 0x1C (“SHR”) opcode.

pub const SAR: Self = _

The 0x1D (“SAR”) opcode.

pub const KECCAK256: Self = _

The 0x20 (“KECCAK256”) opcode.

pub const ADDRESS: Self = _

The 0x30 (“ADDRESS”) opcode.

pub const BALANCE: Self = _

The 0x31 (“BALANCE”) opcode.

pub const ORIGIN: Self = _

The 0x32 (“ORIGIN”) opcode.

pub const CALLER: Self = _

The 0x33 (“CALLER”) opcode.

pub const CALLVALUE: Self = _

The 0x34 (“CALLVALUE”) opcode.

pub const CALLDATALOAD: Self = _

The 0x35 (“CALLDATALOAD”) opcode.

pub const CALLDATASIZE: Self = _

The 0x36 (“CALLDATASIZE”) opcode.

pub const CALLDATACOPY: Self = _

The 0x37 (“CALLDATACOPY”) opcode.

pub const CODESIZE: Self = _

The 0x38 (“CODESIZE”) opcode.

pub const CODECOPY: Self = _

The 0x39 (“CODECOPY”) opcode.

pub const GASPRICE: Self = _

The 0x3A (“GASPRICE”) opcode.

pub const EXTCODESIZE: Self = _

The 0x3B (“EXTCODESIZE”) opcode.

pub const EXTCODECOPY: Self = _

The 0x3C (“EXTCODECOPY”) opcode.

pub const RETURNDATASIZE: Self = _

The 0x3D (“RETURNDATASIZE”) opcode.

pub const RETURNDATACOPY: Self = _

The 0x3E (“RETURNDATACOPY”) opcode.

pub const EXTCODEHASH: Self = _

The 0x3F (“EXTCODEHASH”) opcode.

pub const BLOCKHASH: Self = _

The 0x40 (“BLOCKHASH”) opcode.

pub const COINBASE: Self = _

The 0x41 (“COINBASE”) opcode.

pub const TIMESTAMP: Self = _

The 0x42 (“TIMESTAMP”) opcode.

pub const NUMBER: Self = _

The 0x43 (“NUMBER”) opcode.

pub const DIFFICULTY: Self = _

The 0x44 (“DIFFICULTY”) opcode.

pub const GASLIMIT: Self = _

The 0x45 (“GASLIMIT”) opcode.

pub const CHAINID: Self = _

The 0x46 (“CHAINID”) opcode.

pub const SELFBALANCE: Self = _

The 0x47 (“SELFBALANCE”) opcode.

pub const BASEFEE: Self = _

The 0x48 (“BASEFEE”) opcode.

pub const BLOBHASH: Self = _

The 0x49 (“BLOBHASH”) opcode.

pub const BLOBBASEFEE: Self = _

The 0x4A (“BLOBBASEFEE”) opcode.

pub const POP: Self = _

The 0x50 (“POP”) opcode.

pub const MLOAD: Self = _

The 0x51 (“MLOAD”) opcode.

pub const MSTORE: Self = _

The 0x52 (“MSTORE”) opcode.

pub const MSTORE8: Self = _

The 0x53 (“MSTORE8”) opcode.

pub const SLOAD: Self = _

The 0x54 (“SLOAD”) opcode.

pub const SSTORE: Self = _

The 0x55 (“SSTORE”) opcode.

pub const JUMP: Self = _

The 0x56 (“JUMP”) opcode.

pub const JUMPI: Self = _

The 0x57 (“JUMPI”) opcode.

pub const PC: Self = _

The 0x58 (“PC”) opcode.

pub const MSIZE: Self = _

The 0x59 (“MSIZE”) opcode.

pub const GAS: Self = _

The 0x5A (“GAS”) opcode.

pub const JUMPDEST: Self = _

The 0x5B (“JUMPDEST”) opcode.

pub const TLOAD: Self = _

The 0x5C (“TLOAD”) opcode.

pub const TSTORE: Self = _

The 0x5D (“TSTORE”) opcode.

pub const MCOPY: Self = _

The 0x5E (“MCOPY”) opcode.

pub const PUSH0: Self = _

The 0x5F (“PUSH0”) opcode.

pub const PUSH1: Self = _

The 0x60 (“PUSH1”) opcode.

pub const PUSH2: Self = _

The 0x61 (“PUSH2”) opcode.

pub const PUSH3: Self = _

The 0x62 (“PUSH3”) opcode.

pub const PUSH4: Self = _

The 0x63 (“PUSH4”) opcode.

pub const PUSH5: Self = _

The 0x64 (“PUSH5”) opcode.

pub const PUSH6: Self = _

The 0x65 (“PUSH6”) opcode.

pub const PUSH7: Self = _

The 0x66 (“PUSH7”) opcode.

pub const PUSH8: Self = _

The 0x67 (“PUSH8”) opcode.

pub const PUSH9: Self = _

The 0x68 (“PUSH9”) opcode.

pub const PUSH10: Self = _

The 0x69 (“PUSH10”) opcode.

pub const PUSH11: Self = _

The 0x6A (“PUSH11”) opcode.

pub const PUSH12: Self = _

The 0x6B (“PUSH12”) opcode.

pub const PUSH13: Self = _

The 0x6C (“PUSH13”) opcode.

pub const PUSH14: Self = _

The 0x6D (“PUSH14”) opcode.

pub const PUSH15: Self = _

The 0x6E (“PUSH15”) opcode.

pub const PUSH16: Self = _

The 0x6F (“PUSH16”) opcode.

pub const PUSH17: Self = _

The 0x70 (“PUSH17”) opcode.

pub const PUSH18: Self = _

The 0x71 (“PUSH18”) opcode.

pub const PUSH19: Self = _

The 0x72 (“PUSH19”) opcode.

pub const PUSH20: Self = _

The 0x73 (“PUSH20”) opcode.

pub const PUSH21: Self = _

The 0x74 (“PUSH21”) opcode.

pub const PUSH22: Self = _

The 0x75 (“PUSH22”) opcode.

pub const PUSH23: Self = _

The 0x76 (“PUSH23”) opcode.

pub const PUSH24: Self = _

The 0x77 (“PUSH24”) opcode.

pub const PUSH25: Self = _

The 0x78 (“PUSH25”) opcode.

pub const PUSH26: Self = _

The 0x79 (“PUSH26”) opcode.

pub const PUSH27: Self = _

The 0x7A (“PUSH27”) opcode.

pub const PUSH28: Self = _

The 0x7B (“PUSH28”) opcode.

pub const PUSH29: Self = _

The 0x7C (“PUSH29”) opcode.

pub const PUSH30: Self = _

The 0x7D (“PUSH30”) opcode.

pub const PUSH31: Self = _

The 0x7E (“PUSH31”) opcode.

pub const PUSH32: Self = _

The 0x7F (“PUSH32”) opcode.

pub const DUP1: Self = _

The 0x80 (“DUP1”) opcode.

pub const DUP2: Self = _

The 0x81 (“DUP2”) opcode.

pub const DUP3: Self = _

The 0x82 (“DUP3”) opcode.

pub const DUP4: Self = _

The 0x83 (“DUP4”) opcode.

pub const DUP5: Self = _

The 0x84 (“DUP5”) opcode.

pub const DUP6: Self = _

The 0x85 (“DUP6”) opcode.

pub const DUP7: Self = _

The 0x86 (“DUP7”) opcode.

pub const DUP8: Self = _

The 0x87 (“DUP8”) opcode.

pub const DUP9: Self = _

The 0x88 (“DUP9”) opcode.

pub const DUP10: Self = _

The 0x89 (“DUP10”) opcode.

pub const DUP11: Self = _

The 0x8A (“DUP11”) opcode.

pub const DUP12: Self = _

The 0x8B (“DUP12”) opcode.

pub const DUP13: Self = _

The 0x8C (“DUP13”) opcode.

pub const DUP14: Self = _

The 0x8D (“DUP14”) opcode.

pub const DUP15: Self = _

The 0x8E (“DUP15”) opcode.

pub const DUP16: Self = _

The 0x8F (“DUP16”) opcode.

pub const SWAP1: Self = _

The 0x90 (“SWAP1”) opcode.

pub const SWAP2: Self = _

The 0x91 (“SWAP2”) opcode.

pub const SWAP3: Self = _

The 0x92 (“SWAP3”) opcode.

pub const SWAP4: Self = _

The 0x93 (“SWAP4”) opcode.

pub const SWAP5: Self = _

The 0x94 (“SWAP5”) opcode.

pub const SWAP6: Self = _

The 0x95 (“SWAP6”) opcode.

pub const SWAP7: Self = _

The 0x96 (“SWAP7”) opcode.

pub const SWAP8: Self = _

The 0x97 (“SWAP8”) opcode.

pub const SWAP9: Self = _

The 0x98 (“SWAP9”) opcode.

pub const SWAP10: Self = _

The 0x99 (“SWAP10”) opcode.

pub const SWAP11: Self = _

The 0x9A (“SWAP11”) opcode.

pub const SWAP12: Self = _

The 0x9B (“SWAP12”) opcode.

pub const SWAP13: Self = _

The 0x9C (“SWAP13”) opcode.

pub const SWAP14: Self = _

The 0x9D (“SWAP14”) opcode.

pub const SWAP15: Self = _

The 0x9E (“SWAP15”) opcode.

pub const SWAP16: Self = _

The 0x9F (“SWAP16”) opcode.

pub const LOG0: Self = _

The 0xA0 (“LOG0”) opcode.

pub const LOG1: Self = _

The 0xA1 (“LOG1”) opcode.

pub const LOG2: Self = _

The 0xA2 (“LOG2”) opcode.

pub const LOG3: Self = _

The 0xA3 (“LOG3”) opcode.

pub const LOG4: Self = _

The 0xA4 (“LOG4”) opcode.

pub const DATALOAD: Self = _

The 0xD0 (“DATALOAD”) opcode.

pub const DATALOADN: Self = _

The 0xD1 (“DATALOADN”) opcode.

pub const DATASIZE: Self = _

The 0xD2 (“DATASIZE”) opcode.

pub const DATACOPY: Self = _

The 0xD3 (“DATACOPY”) opcode.

pub const RJUMP: Self = _

The 0xE0 (“RJUMP”) opcode.

pub const RJUMPI: Self = _

The 0xE1 (“RJUMPI”) opcode.

pub const RJUMPV: Self = _

The 0xE2 (“RJUMPV”) opcode.

pub const CALLF: Self = _

The 0xE3 (“CALLF”) opcode.

pub const RETF: Self = _

The 0xE4 (“RETF”) opcode.

pub const JUMPF: Self = _

The 0xE5 (“JUMPF”) opcode.

pub const DUPN: Self = _

The 0xE6 (“DUPN”) opcode.

pub const SWAPN: Self = _

The 0xE7 (“SWAPN”) opcode.

pub const EXCHANGE: Self = _

The 0xE8 (“EXCHANGE”) opcode.

pub const EOFCREATE: Self = _

The 0xEC (“EOFCREATE”) opcode.

pub const RETURNCONTRACT: Self = _

The 0xEE (“RETURNCONTRACT”) opcode.

pub const CREATE: Self = _

The 0xF0 (“CREATE”) opcode.

pub const CALL: Self = _

The 0xF1 (“CALL”) opcode.

pub const CALLCODE: Self = _

The 0xF2 (“CALLCODE”) opcode.

pub const RETURN: Self = _

The 0xF3 (“RETURN”) opcode.

pub const DELEGATECALL: Self = _

The 0xF4 (“DELEGATECALL”) opcode.

pub const CREATE2: Self = _

The 0xF5 (“CREATE2”) opcode.

pub const RETURNDATALOAD: Self = _

The 0xF7 (“RETURNDATALOAD”) opcode.

pub const EXTCALL: Self = _

The 0xF8 (“EXTCALL”) opcode.

pub const EXTDELEGATECALL: Self = _

The 0xF9 (“EXTDELEGATECALL”) opcode.

pub const STATICCALL: Self = _

The 0xFA (“STATICCALL”) opcode.

pub const EXTSTATICCALL: Self = _

The 0xFB (“EXTSTATICCALL”) opcode.

pub const REVERT: Self = _

The 0xFD (“REVERT”) opcode.

pub const INVALID: Self = _

The 0xFE (“INVALID”) opcode.

pub const SELFDESTRUCT: Self = _

The 0xFF (“SELFDESTRUCT”) opcode.