Trait revm_interpreter::Host

pub trait Host {

    // Required methods
    fn env(&mut self) -> &mut Env;
    fn load_account(&mut self, address: Address) -> Option<(bool, bool)>;
    fn block_hash(&mut self, number: U256) -> Option<B256>;
    fn balance(&mut self, address: Address) -> Option<(U256, bool)>;
    fn code(&mut self, address: Address) -> Option<(Bytecode, bool)>;
    fn code_hash(&mut self, address: Address) -> Option<(B256, bool)>;
    fn sload(&mut self, address: Address, index: U256) -> Option<(U256, bool)>;
    fn sstore(
        &mut self,
        address: Address,
        index: U256,
        value: U256
    ) -> Option<(U256, U256, U256, bool)>;
    fn tload(&mut self, address: Address, index: U256) -> U256;
    fn tstore(&mut self, address: Address, index: U256, value: U256);
    fn log(&mut self, address: Address, topics: Vec<B256>, data: Bytes);
    fn call(
        &mut self,
        input: &mut CallInputs,
        shared_memory: &mut SharedMemory
    ) -> (InstructionResult, Gas, Bytes);
    fn create(
        &mut self,
        inputs: &mut CreateInputs,
        shared_memory: &mut SharedMemory
    ) -> (InstructionResult, Option<Address>, Gas, Bytes);
    fn selfdestruct(
        &mut self,
        address: Address,
        target: Address
    ) -> Option<SelfDestructResult>;

}

EVM context host.

Required Methods

fn env(&mut self) -> &mut Env

Returns a mutable reference to the environment.

fn load_account(&mut self, address: Address) -> Option<(bool, bool)>

Load an account.

Returns (is_cold, is_new_account)

fn block_hash(&mut self, number: U256) -> Option<B256>

Get the block hash of the given block number.

fn balance(&mut self, address: Address) -> Option<(U256, bool)>

Get balance of address and if the account is cold.

fn code(&mut self, address: Address) -> Option<(Bytecode, bool)>

Get code of address and if the account is cold.

fn code_hash(&mut self, address: Address) -> Option<(B256, bool)>

Get code hash of address and if the account is cold.

fn sload(&mut self, address: Address, index: U256) -> Option<(U256, bool)>

Get storage value of address at index and if the account is cold.

fn sstore( &mut self, address: Address, index: U256, value: U256 ) -> Option<(U256, U256, U256, bool)>

Set storage value of account address at index.

Returns (original, present, new, is_cold).

fn tload(&mut self, address: Address, index: U256) -> U256

Get the transient storage value of address at index.

fn tstore(&mut self, address: Address, index: U256, value: U256)

Set the transient storage value of address at index.

fn log(&mut self, address: Address, topics: Vec<B256>, data: Bytes)

Emit a log owned by address with given topics and data.

fn call( &mut self, input: &mut CallInputs, shared_memory: &mut SharedMemory ) -> (InstructionResult, Gas, Bytes)

Invoke a call operation.

fn create( &mut self, inputs: &mut CreateInputs, shared_memory: &mut SharedMemory ) -> (InstructionResult, Option<Address>, Gas, Bytes)

Invoke a create operation.

fn selfdestruct( &mut self, address: Address, target: Address ) -> Option<SelfDestructResult>

Mark address to be deleted, with funds transferred to target.