revm_interpreter::instructions::utility

Function cast_slice_to_u256

pub fn cast_slice_to_u256(slice: &[u8], dest: &mut U256)

Pushes an arbitrary length slice of bytes onto the stack, padding the last word with zeros if necessary.

Panics

Panics if slice is longer than 32 bytes.